Refactor components to TypeScript
