+++ client/package.json
@@ -52,6 +52,7 @@
         "@types/qs": "^6.14.0",
         "@types/react": "^19.2.13",
         "@types/react-dom": "^19.2.3",
+        "@types/react-responsive-masonry": "^2.7.0",
         "eslint": "^9.23.0",
         "eslint-formatter-junit": "^9.0.0",
         "eslint-plugin-jsdoc": "^62.0.0",
