+++ client/src/component/tag/TagList.tsx
@@ -2,11 +2,17 @@
  * Colly | tag list
  */
 
+import type { TagRes } from "./../../logic/api/tag"
 import TagBadge from "./TagBadge"
 
 import "./TagList.css"
 
-function TagList({ tags, clickAction }) {
+type TagListProps = {
+    tags: TagRes[]
+    clickAction?: (tag: TagRes) => void
+}
+
+function TagList({ tags, clickAction }: TagListProps) {
     return (
         <div className="tags-spacing">
             {tags.map((tag) => (
