+++ client/src/main.tsx
@@ -7,7 +7,7 @@ import ReactDOM from "react-dom/client"
 
 import App from "./component/App"
 
-const root = ReactDOM.createRoot(document.getElementById("root"))
+const root = ReactDOM.createRoot(document.getElementById("root")!)
 root.render(
     <React.StrictMode>
         <App />
